Hit record simplification and removal
Removed the hittable and hit record classes.

Really unnecessary at this point as we're focused on only hitting triangles and boxes, not a slew of arbitrary parametric shapes.

Updated hit test functions as well, due to previous dependence on intervals and hit records. T-value and 'record' are now just stored and updated in the ray itself.

Model also now no longer has a 'hit' function, the bvh is publicly accessible, so it's tested directly. This may change if we move to a per model ray transform method for testing.