First pass conversion to GLM
Moving to GLM as work starts on moving to GPU and integration of Vulkan. Having immediate access to built-in matrix math and other helpers will be paramount.

Though GLM is still missing some helpers that need to be kept around and there is still lots of cleaning up to do. vec3.h remains in as classes still reference it for such helpers.